requirements.txt

streamlit
langchain
langchain-openai
openai
python-dotenv
pandas
pytest
pytest-cov


